* More work on the algorithm, not good though

+++ trunk/ModelBuilder.h
@@ -29,7 +29,7 @@ private:
 	// vector of points
 	std::vector<glm::uvec3> points;
 	// vector of quads to hold the connectivity
-	std::vector<glm::vec4> quads;
+	std::vector<glm::uvec4> quads;
 	/* build polygons recursively based on nearby points. Current index referes to the current pixel to be processed, 0 if it's the first. */
 	void BuildPolygon(glm::uvec2 size, short* depthBuffer);
 };
